+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 - VLS grating parameter calculations (beta)
 ### Changed
-
+- add source bandwidth to results of raypyng analysis
+- Normalize flux to 0.1% bandwidth
 ### Fixed